nstead of setting the pages and titles list every time the method is run, it now only sets the title and pages once if the device is found. Also, I removed the key parameter for the MyApp and RootPage widgets, since it is not being used.
I have also changed the pages variable to a non-const variable, and the initialization of the pages list in the initState method, so that it can be reassigned in runFunctionOnStartup.

Also, the titles list that is not used anymore, I have replaced it with a single string variable named title.

Finally, I have removed the commented out code, as it is not needed.